Move and Rename FetchBillingAccountsService
