+++ app/services/bill-runs/two-part-tariff/generate-bill-run.service.js
@@ -14,7 +14,7 @@ const {
   currentTimeInNanoseconds,
   timestampForPostgres
 } = require('../../../lib/general.lib.js')
-const FetchBillingAccountsService = require('./fetch-billing-accounts.service.js')
+const FetchTwoPartTariffBillingAccountsService = require('../fetch-two-part-tariff-billing-accounts.service.js')
 const HandleErroredBillRunService = require('../handle-errored-bill-run.service.js')
 const LegacyRefreshBillRunRequest = require('../../../requests/legacy/refresh-bill-run.request.js')
 const ProcessBillingPeriodService = require('./process-billing-period.service.js')
@@ -77,7 +77,7 @@ function _billingPeriod(billRun) {
 
 async function _fetchBillingAccounts(billRunId) {
   try {
-    return await FetchBillingAccountsService.go(billRunId)
+    return await FetchTwoPartTariffBillingAccountsService.go(billRunId)
   } catch (error) {
     // We know we're saying we failed to process charge versions. But we're stuck with the legacy error codes and this
     // is the closest one related to what stage we're at in the process
