+++ test/services/bill-runs/fetch-two-part-tariff-billing-accounts.service.test.js
@@ -8,24 +8,24 @@ const { describe, it, before } = (exports.lab = Lab.script())
 const { expect } = Code
 
 // Test helpers
-const BillRunHelper = require('../../../support/helpers/bill-run.helper.js')
-const BillingAccountHelper = require('../../../support/helpers/billing-account.helper.js')
-const BillingAccountModel = require('../../../../app/models/billing-account.model.js')
-const ChargeCategoryHelper = require('../../../support/helpers/charge-category.helper.js')
-const ChargeElementHelper = require('../../../support/helpers/charge-element.helper.js')
-const ChargeReferenceHelper = require('../../../support/helpers/charge-reference.helper.js')
-const ChargeVersionHelper = require('../../../support/helpers/charge-version.helper.js')
-const LicenceHelper = require('../../../support/helpers/licence.helper.js')
-const RegionHelper = require('../../../support/helpers/region.helper.js')
-const ReviewChargeElementHelper = require('../../../support/helpers/review-charge-element.helper.js')
-const ReviewChargeReferenceHelper = require('../../../support/helpers/review-charge-reference.helper.js')
-const ReviewChargeVersionHelper = require('../../../support/helpers/review-charge-version.helper.js')
-const ReviewLicenceHelper = require('../../../support/helpers/review-licence.helper.js')
+const BillRunHelper = require('../../support/helpers/bill-run.helper.js')
+const BillingAccountHelper = require('../../support/helpers/billing-account.helper.js')
+const BillingAccountModel = require('../../../app/models/billing-account.model.js')
+const ChargeCategoryHelper = require('../../support/helpers/charge-category.helper.js')
+const ChargeElementHelper = require('../../support/helpers/charge-element.helper.js')
+const ChargeReferenceHelper = require('../../support/helpers/charge-reference.helper.js')
+const ChargeVersionHelper = require('../../support/helpers/charge-version.helper.js')
+const LicenceHelper = require('../../support/helpers/licence.helper.js')
+const RegionHelper = require('../../support/helpers/region.helper.js')
+const ReviewChargeElementHelper = require('../../support/helpers/review-charge-element.helper.js')
+const ReviewChargeReferenceHelper = require('../../support/helpers/review-charge-reference.helper.js')
+const ReviewChargeVersionHelper = require('../../support/helpers/review-charge-version.helper.js')
+const ReviewLicenceHelper = require('../../support/helpers/review-licence.helper.js')
 
 // Thing under test
-const FetchBillingAccountsService = require('../../../../app/services/bill-runs/two-part-tariff/fetch-billing-accounts.service.js')
+const FetchTwoPartTariffBillingAccountsService = require('../../../app/services/bill-runs/fetch-two-part-tariff-billing-accounts.service.js')
 
-describe('Fetch Billing Accounts service', () => {
+describe('Bill Runs - Fetch Two Part Tariff Billing Accounts service', () => {
   let billRun
   let billingAccount
   let billingAccountNotInBillRun
@@ -79,7 +79,7 @@ describe('Fetch Billing Accounts service', () => {
 
   describe('when there are billing accounts that are linked to a two-part tariff bill run', () => {
     it('returns the applicable billing accounts', async () => {
-      const results = await FetchBillingAccountsService.go(billRun.id)
+      const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
       expect(results).to.have.length(1)
 
@@ -91,7 +91,7 @@ describe('Fetch Billing Accounts service', () => {
     describe('and each billing account', () => {
       describe('for the charge versions property', () => {
         it('returns the applicable charge versions', async () => {
-          const results = await FetchBillingAccountsService.go(billRun.id)
+          const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
           const { chargeVersions } = results[0]
 
@@ -105,7 +105,7 @@ describe('Fetch Billing Accounts service', () => {
 
         describe('and against each charge version', () => {
           it('includes the licence', async () => {
-            const results = await FetchBillingAccountsService.go(billRun.id)
+            const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
             const { licence } = results[0].chargeVersions[0]
 
@@ -121,7 +121,7 @@ describe('Fetch Billing Accounts service', () => {
           })
 
           it('includes the applicable charge references', async () => {
-            const results = await FetchBillingAccountsService.go(billRun.id)
+            const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
             const { chargeReferences } = results[0].chargeVersions[0]
 
@@ -143,7 +143,7 @@ describe('Fetch Billing Accounts service', () => {
 
           describe('and against each charge reference', () => {
             it('includes the charge category', async () => {
-              const results = await FetchBillingAccountsService.go(billRun.id)
+              const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
               const { chargeCategory: result } = results[0].chargeVersions[0].chargeReferences[0]
 
@@ -153,7 +153,7 @@ describe('Fetch Billing Accounts service', () => {
             })
 
             it('includes the review charge references', async () => {
-              const results = await FetchBillingAccountsService.go(billRun.id)
+              const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
               const { reviewChargeReferences: result } = results[0].chargeVersions[0].chargeReferences[0]
 
@@ -164,7 +164,7 @@ describe('Fetch Billing Accounts service', () => {
             })
 
             it('includes the charge elements', async () => {
-              const results = await FetchBillingAccountsService.go(billRun.id)
+              const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
               const { chargeElements: result } = results[0].chargeVersions[0].chargeReferences[0]
 
@@ -177,7 +177,7 @@ describe('Fetch Billing Accounts service', () => {
 
             describe('and against each charge element', () => {
               it('includes the review charge elements', async () => {
-                const results = await FetchBillingAccountsService.go(billRun.id)
+                const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
                 const { reviewChargeElements: result } =
                   results[0].chargeVersions[0].chargeReferences[0].chargeElements[0]
@@ -194,7 +194,7 @@ describe('Fetch Billing Accounts service', () => {
 
   describe('when there are billing accounts not linked to a two-part tariff bill run', () => {
     it('does not include them in the results', async () => {
-      const results = await FetchBillingAccountsService.go(billRun.id)
+      const results = await FetchTwoPartTariffBillingAccountsService.go(billRun.id)
 
       expect(results).to.have.length(1)
 
@@ -205,7 +205,7 @@ describe('Fetch Billing Accounts service', () => {
 
   describe('when there are no billing accounts at all (no results)', () => {
     it('returns no results', async () => {
-      const results = await FetchBillingAccountsService.go('1c1f7af5-9cba-47a7-8fc4-2c03b0d1124d')
+      const results = await FetchTwoPartTariffBillingAccountsService.go('1c1f7af5-9cba-47a7-8fc4-2c03b0d1124d')
 
       expect(results).to.be.empty()
     })
